Move the page "/other" components to a folder.
