+++ src/routes/other/components/DeleteAllDataBlock.svelte
